Add config file parsing
This adds a config package which support parsing a yaml file as
configuration to provide customization at-runtime.

For now the only real configurable piece is the HTTP port, but in the
future this will be useful for many other things like database
connection details, feature flags, storage backend configs, etc.