Add and configure gorm and postgres
This adds database support, somewhat, by introducing GORM and PostgreSQL
(the latter with a docker container via docker-compose), configured via
the new config.yaml system.

This does not store the DB connection anywhere yet, and nothing actually
uses the database connection yet, so there is still work to be done for
this to be functional. However, this does get the foundational stuff in
place so we're reading to move forward.