Introduce data package and user/auth models
To isolate database access, we introduce a subpackage just for database
things, which will allow us to avoid passing around the database
connection pool everywhere.

We also introduce our first two models, as part of the first attempts to
get Nextcloud "Login V2" authentication working. Since the Nextcloud
auth flow is kind of weird, we're opting to call it "nextcloud"
specifically so we are free to invent something better down the road
when we have time to implement our own clients.

The user model also forgoes any password or other local authentication
method for now, just to keep things simple as building a full
authentication system is not the priority at the moment.