Implement first draft of auth/login system
This is a big spike of work that implements a very (very) basic login
and session system, as well as the Nextcloud API endpoints.

It's still a bit messy, and probably has many bugs, but it does work. I
skipped actually storing credentials on a User, not for any reason other
than I thought it would save some work. Ultimately I had to pull bcrypt
in and deal with password-like values for the NextcloudAppPasswords
anyway, so it was mostly for nothing. We'll get there eventually. For
now I'm more concerned about getting the core behaviour and file syncing
built out.

With this, we can successfully add an account to the Nextcloud desktop
app, though it errors immediately since we haven't implemented, you
know, the rest of the API. However, it's a good step forward!