Add models for file index package
We're going to need to model files and directories in the database to
keep an index of the known files and directories for each user. This
will be the primary datasource for listing files, looking up metadata,
permissions, etc.

These are not used yet as of this commit, but will be shortly as we
build out a filesystem indexer and storage backend api.