Add basic file indexing and local storage backend
This spikes out the beginnings of a way to store files on-disk, through
the FilesystemBackend implementation of the new StorageBackend
interface; we also add a crawler to iterate through all the files and
directories in the StorageBackend to discover them and add them in the
database.

Starting just with a local filesystem backend is the easiest and fastest
path to success for now. Eventually more exotic backends like object
storage would be great, but for now we just want to get something simple
working.

We make a goroutine for each directory to process multiple in parallel.
With my small test data locally that improves the scan time by over 30%,
and there's probably more we can squeeze out of it in the future.