Add ID as prefix to AppPasswords
Because we're hashing the app password values, we can't actually look
them up anymore because they'll be salted so the values won't match.

To solve this, we'll create a "composite" password syntax where the ID
is prefixed with `-` to the front. We can then pull this off and use
that to find the AppPassword row so that we can use the proper bcrypt
comparison function with the password digest.