Only serialise requested properties for a PROPFIND
As part of webdav, we need to support only serialising the properties
that were asked for, instead of barfing out all of them every time.

This gives us a chance to refactor some of the property handling code.
We now have a map of requested property names to a function that can
return a `DavProp` struct for that property with the right value(s).

In the future, we'll want to abstract away a lot of the
nextcloud-specific property handling. It would be nice to expose an API
to allow injecting custom property handlers from the caller so we can
isolate the nextcloud-specific properties in the nextcloud package.