Add a static OCS "capabilities" endpoint
This returns some fixed "capabilities" data to pretend that we're a
Nextcloud 22 server that supports theming and webdav chunking.

Note that as of this commit we _do not_ support chunking, though
changing the value returned here does not actually affect the client as
it will send us chunks regardless.

What this _does_ change, however, is by returning `dav.chunking = 1.0`,
we make the clients switch to "next gen" chunking, which stores the
chunks at a different path (under `uploads`) which will make it easier
for us to intercept and handle these requests.