Spike out upload chunking support
This adds three new webdav endpoints to support Nextcloud's "next-gen
chunking" API, which uploads files as a series of variable-size chunks
and then has the server reconstruct that file and move it to the right
destination when all the chunks are uploaded.

There's a *lot* we can clean up here, but this at least makes it work.