Add hacky avatar endpoint
Kind of a fun one to proxy requests to the avatars to Gravatar as a
quick and dirty way to get working avatars without implementing our own
system.

This should eventually be replaced by our own in-house avatar system,
which could still default to gravatar, but should be served from our own
cache/storage.