Lower AppPassword hash cost to balance performance
This was the result of my complete oversight of the performance cost of
a high bcrypt cost. Since AppPasswords have to be authenticated for
every request, that meant an extremely costly computation for *every
HTTP request*. With 10 rounds, this meant 100+ms of time burned just on
bcrypt alone per request.

Lowering this cost "solves" the performace problem at the expense of
some security. In reality, the industry standard for this sort of token
is to just store them raw, which we are still kilometers ahead of, even
with a low cost.

Realistically, the threat model here is just to protect against
undesired access in the face of database compromise. Since App Passwords
are easy to rotate (just truncate the whole table and force everyone to
log in again), the risk of having weaker hashes is mitigated by the fact
we can just blow away the entire table whenever we want.

This seems like a fair compromise to me, as we shave nearly 100ms off of
every request at very little expense in security. It's still enough
bcrypt to slow a hacker down from gaining immediate API access, but not
so much hashing that we destroy our own servers and experience in the
process.