Add CSRF protection
This adds a tiny session-based CSRF protection package to generate
tokens for each request and validate the one from the form with the one
in the session it generated.

This should be sufficient for most things, though we may want to
generalize the storage of CSRF tokens in the future.