Add flag to run seeds in Go
Instead of the haphazard sql file to insert seed data, this moves it
under a CLI flag and to use the normal Go-based functions we have.

This standardizes how the data is inserted, and lets us do things like
foreign keys and timestamps much easier as it is all handled
automatically properly by the ORM instead of us hacking it in there.