Add web package, move login page to it
As we think about how we're going to provide a basic web UI in the MVP,
the path of least pain is likely just to continue with server-side
rendering. As such, we should probably start isolating the "web UI"
parts in their own package to isolate the "web app" from the rest of the
app.

This will make it easier to replace later, if we need to move to a more
JS-focused frontend, as all the routes and handlers we need to replace
will be centralised under a single namespace.

Things like the nextcloud login flow will stay where they are for now,
as they aren't really part of the "nucleus web app". We can probably
generalise them to be part of our web frontend instead of their own
little thing, but that's a future refactor and not a priority right now.