Add support for renaming files
We need to be able to move a file from one path to another. This
presents a bit of a challenge, as for a directory we need to effect this
change down the tree to make sure the projected FullName on all the
children is updated.

This is a bit of tradeoff, as having the FullName makes lookup queries
much lighter and easier, but does make renames a bit of a heavy
operation. Since reads are significantly more likely than renames, this
seems like a fair compromise.

There may be (/are probably) ways we can make this less annoying and
less thrashy, but for now this does solve it and should remain
performant for a while.