Clean up main, move routes into router
This tries to simplify what we put in `main()` by splitting the routes
and Gin stuff into its own file/methods, and by moving the flags into
"global" vars.

Really `main()` should only be for parsing the flags and handling the
topmost control flow of the program. With these latest changes it seems
like we're basically at that point.