Change nxc CurrentUser to middleware
This aligns the nxc CurrentUser function to match how we refactored it
in the standard auth middleware. By adding this as middleware and
storing the user in the gin.Context we can assume it exists and avoid
some error checks in the handlers themselves.

If we ever add tests for these gin handlers, this will also make them
significantly easier to test as we can just inject a user into the
context. This indicates to me that this context-based approach is a much
better way to handle this.