Delete temporary upload files after completion
Once a chunked upload completes, we're currently leaving the file chunks
around in the backend. This will eventually become an annoying problem,
and it'd be nice if we didn't need infrastructure in place to deal with
this and the program could clean up its mess automatically.

We should just add a small function to delete an upload path and call
this at the end of the final MOVE after a chunked upload completes, as
we know by that point that we can safely remove all the temporary files.