Remove FileDigest from backend interface
This is the wrong place for this logic, it has nothing to do with
storing files, it's just a utility function.

I'm not 100% happy with now duplicating the sha1 hash code in two
different places, but I think this is still better than committing to
having an interface function that makes no sense.

This might be a red flag that we're missing an abstraction between the
callers and the FilesystemBackend.