Add very basic user management UI
We need to be able to add and remove users from the system. This
implements the absolute minimum of that functionality.

This whole view should eventually be limited to only "admin" users, but
that concept hasn't been implemented yet so for now this is visible to
any user of the system.

Deleting a user with files fails due to the foreign key constraint,
which is a nice accidental validation that prevents deleting an active
user.