Move user settings to /me
While the user settings are the only real thing we provide at the
moment, it still doesn't make a lot of sense for them to be at the root
of the web app.

Let's make an empty dashboard handler to redirect to the user settings,
so it's less disruptive later when we eventually add a dashboard.