Add NextcloudAppPasswords to account settings
This surfaces NextcloudAppPasswords on the settings page in a table, and
allows a user to revoke them individually. This is a critical security
feature.

Unfortunately there's not much to differentiate them from one another
except the creation timestamp... Eventually it'd be nice to record some
more identifiable information such as OS platform, IP country, etc.