Add Nextcloud login v1 flow support
I thought the V2 login flow was sufficient, but it seems the mobile apps
(or iOS, at least) still require login v1. This is a very simple flow as
it's just a single step, so we just reimplement the poll session in a
single go and return the AppPassword at the end as the password.