Support PROPFIND without trailing slash
The iOS app appears to sometimes (not consistently lol) make PROPFIND
requests for the root directory without a trailing slash. This is
different from the desktop client which consistently requests with a
trailing slash.

It seems reasonable to make the trailing slash optional, so let's just
add another handler to allow that. There might be a fancy way to make
the filePath optional in the other Gin handler, but for now this works.