Move route definitions for webdav to webdav router
The coupling between the nxc router and webdav router was a bit ... odd.
We were only one step away from making the webdav router
self-sufficient, but for whatever reason decided to put all the route
definitions in the nxc router directly...

We should just move the route definitions into the webdav router and
make it a full, proper router.