Limit number of active indexers in Crawler
Currently we just spin off as many goroutines as we desire (one per
folder found) with reckless abandon. This can lead to some pretty severe
resource contention on very large file trees, or on lower power
computers.

We should introduce a counting semaphore to ensure that we block all the
goroutines beyond a reasonable limit. This should not adversely affect
performance, and may even improve it by throttling on our side instead
of relying on the OS IO to throttle us.